requirements.txt
Flask
Flask-SQLAlchemy
pytesseract
Pillow
gunicorn
